refactor(fleet-core): move the AI gateway HTTP router into core-ai-gateway
