+++ packages/core-ai-gateway/src/gateway-router/types.ts
@@ -1,0 +1,11 @@
+import type http from "node:http";
+
+export type GatewayHttpHandlerContext = {
+  readonly req: http.IncomingMessage;
+  readonly res: http.ServerResponse;
+  readonly pathname: string;
+};
+
+export type GatewayHttpHandler = (
+  ctx: GatewayHttpHandlerContext,
+) => boolean | Promise<boolean>;
