+++ packages/core-ai-gateway/src/index.ts
@@ -3,6 +3,11 @@
 export * from "./anthropic/index.js";
 export * from "./canonical/index.js";
 export * from "./cursor/index.js";
+export * from "./cursor/diagnostic-log.js";
+export * from "./gateway-router/router.js";
+export * from "./gateway-router/types.js";
+export * from "./gateway-router/proxy.js";
+export * from "./gateway-router/opencode.js";
 export * from "./transport/credentials.js";
 // wireLog/logCanonicalEvents 는 진단 구현 세부사항이므로 런타임 제어 표면만 좁게 공개한다.
 export { DEFAULT_WIRE_LOG_MAX_BYTES, setWireLogTarget, wireLogEnabled } from "./transport/wire-log.js";
